+++ tests/edit/set-version/14.sh
@@ -1,0 +1,1 @@
+${PORTEDIT} set-version 1.1 14.in | diff -L 14.expected -L 14.actual -u 14.expected -